OGM-1002 Move integration tests in the same project
  * Remove the modules from the pom and make it non deployable jar
  * Create a package for the test cases
  * Create new packages for Cassandra, Infinispan, EhCache
