+++ integrationtest/src/test/resources/infinispan.xml
