+++ integrationtest/src/test/resources/jboss-deployment-structure-ehcache.xml
