+++ integrationtest/src/test/resources/module-versions.properties
