--- integrationtest/testcase/pom.xml
@@ -1,146 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>org.hibernate.ogm</groupId>
-        <artifactId>hibernate-ogm-integrationtest</artifactId>
-        <version>5.0.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>hibernate-ogm-integrationtest-testcase</artifactId>
-    <name>Hibernate OGM Integration Test case</name>
-    <description>Hibernate OGM Integration test case</description>
-
-    <build>
-        <testResources>
-            <testResource>
-                <directory>src/test/resources</directory>
-                <filtering>true</filtering>
-            </testResource>
-        </testResources>
-        <plugins>
-            <plugin>
-                <artifactId>maven-checkstyle-plugin</artifactId>
-            </plugin>
-            <plugin>
-                <artifactId>maven-deploy-plugin</artifactId>
-            </plugin>
-            <plugin>
-                <artifactId>maven-dependency-plugin</artifactId>
-            </plugin>
-            <plugin>
-                <artifactId>maven-jar-plugin</artifactId>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-failsafe-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>generic-integration-tests</id>
-                        <goals>
-                            <goal>integration-test</goal>
-                        </goals>
-                        <configuration>
-                            <forkMode>once</forkMode>
-                            <includes>
-                                <include>**/*IT.java</include>
-                            </includes>
-                            <redirectTestOutputToFile>true</redirectTestOutputToFile>
-                            <summaryFile>target/failsafe-reports/failsafe-generic-tests.xml</summaryFile>
-                        </configuration>
-                    </execution>
-                    <!--
-                         Now merge all reports from all activated profiles:
-                         There must be a single 'verify' step to merge all results, so these can't be added by profiles.
-                         See also "Running integration tests multiple times" at : http://maven.apache.org/surefire/maven-failsafe-plugin/usage.html
-                    -->
-                    <execution>
-                        <id>verify</id>
-                        <goals>
-                            <goal>verify</goal>
-                        </goals>
-                        <configuration>
-                            <summaryFiles>
-                                <summaryFile>target/failsafe-reports/failsafe-generic-tests.xml</summaryFile>
-                                <summaryFile>target/failsafe-reports/failsafe-cassandra-tests.xml</summaryFile>
-                            </summaryFiles>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-
-    <profiles>
-        <profile>
-            <id>enable-cassandra-integration-tests</id>
-            <!-- To run the tests you need to define the environment property: CASSANDRA_HOSTNAME -->
-            <activation>
-                <property>
-                    <name>env.CASSANDRA_HOSTNAME</name>
-                </property>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-failsafe-plugin</artifactId>
-                        <executions>
-                            <execution>
-                                <id>cassandra-integration-tests</id>
-                                <goals>
-                                    <goal>integration-test</goal>
-                                </goals>
-                                <configuration>
-                                    <includes>
-                                        <include>**/*ITCassandra.java</include>
-                                    </includes>
-                                    <environmentVariables>
-                                        <CASSANDRA_HOSTNAME>${env.CASSANDRA_HOSTNAME}</CASSANDRA_HOSTNAME>
-                                        <CASSANDRA_PORT>${env.CASSANDRA_PORT}</CASSANDRA_PORT>
-                                    </environmentVariables>
-                                    <forkMode>once</forkMode>
-                                    <redirectTestOutputToFile>true</redirectTestOutputToFile>
-                                    <summaryFile>target/failsafe-reports/failsafe-cassandra-tests.xml</summaryFile>
-                                </configuration>
-                            </execution>
-                        </executions>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-        <profile>
-            <!-- Even if Cassandra integration tests are disabled,
-            we still need to generate a valid (empty) report file to be inspected by the verify phase -->
-            <id>disabled-cassandra-integration-tests</id>
-            <activation>
-                <property>
-                    <name>!env.CASSANDRA_HOSTNAME</name>
-                </property>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-failsafe-plugin</artifactId>
-                        <executions>
-                            <execution>
-                                <id>cassandra-integration-tests</id>
-                                <goals>
-                                    <goal>integration-test</goal>
-                                </goals>
-                                <configuration>
-                                    <excludes>
-                                        <exclude>**/*</exclude>
-                                    </excludes>
-                                    <forkMode>none</forkMode>
-                                    <summaryFile>target/failsafe-reports/failsafe-cassandra-tests.xml</summaryFile>
-                                </configuration>
-                            </execution>
-                        </executions>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-    </profiles>
-
-</project>
