--- integrationtest/testcase/src/test/java/org/hibernate/ogm/test/integration/jboss/model/Address.java
@@ -1,48 +1,0 @@
-/*
- * Hibernate OGM, Domain model persistence for NoSQL datastores
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.ogm.test.integration.jboss.model;
-
-import javax.persistence.Embeddable;
-
-/**
- * @author Gunnar Morling
- */
-@Embeddable
-public class Address {
-
-	private String street;
-	private String city;
-
-	Address() {
-	}
-
-	public Address(String street, String city) {
-		this.street = street;
-		this.city = city;
-	}
-
-	public String getStreet() {
-		return street;
-	}
-
-	public void setStreet(String street) {
-		this.street = street;
-	}
-
-	public String getCity() {
-		return city;
-	}
-
-	public void setCity(String city) {
-		this.city = city;
-	}
-
-	@Override
-	public String toString() {
-		return "Address [street=" + street + ", city=" + city + "]";
-	}
-}
