--- integrationtest/testcase/src/test/java/org/hibernate/ogm/test/integration/jboss/model/MagicCard.java
@@ -1,104 +1,0 @@
-/*
- * Hibernate OGM, Domain model persistence for NoSQL datastores
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.ogm.test.integration.jboss.model;
-
-import java.util.Date;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
-
-import org.hibernate.search.annotations.Analyze;
-import org.hibernate.search.annotations.Field;
-import org.hibernate.search.annotations.Indexed;
-
-@Entity
-@Indexed
-public class MagicCard {
-
-	@Id
-	@GeneratedValue
-	private Long id;
-
-	@Field(analyze = Analyze.NO)
-	private String name;
-
-	@Field
-	private String artist;
-
-	@Field
-	private String manacost;
-
-	@Field
-	private Integer power;
-
-	@Field
-	private Integer thoughness;
-
-	@Field
-	@Temporal(TemporalType.DATE)
-	private Date publicationDate;
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public String getArtist() {
-		return artist;
-	}
-
-	public void setArtist(String artist) {
-		this.artist = artist;
-	}
-
-	public String getManacost() {
-		return manacost;
-	}
-
-	public void setManacost(String manacost) {
-		this.manacost = manacost;
-	}
-
-	public Integer getPower() {
-		return power;
-	}
-
-	public void setPower(Integer power) {
-		this.power = power;
-	}
-
-	public Integer getThoughness() {
-		return thoughness;
-	}
-
-	public void setThoughness(Integer thoughness) {
-		this.thoughness = thoughness;
-	}
-
-	public Date getPublicationDate() {
-		return publicationDate;
-	}
-
-	public void setPublicationDate(Date publicationDate) {
-		this.publicationDate = publicationDate;
-	}
-
-}
